Test 'sanitize=fuzzer' before building test/fuzz
We unfortunately can't rely on 'llvm_map_components_to_libnames'
to test for the existence of the  clang fuzzing library since older
versions of the CMake LLVM package (~3.10) will throw fatal errors
when 'llvm_map_components_to_libnames' fails. Though fixed in newer
versions, we can't support 'llvm_map_components_to_libnames' on
some of our older platforms. The next best solution is to check if
we can compile a test program with the '-fsanitize=fuzzer' linker
flag. If this succeeds, we should be able to compile 'test/fuzz'